Refactor classifer utils.
