+++ .gitignore
@@ -11,7 +11,7 @@ __pycache__/
 .vscode
 
 # Keras Models
-model/*
+*model/*
 
 # Distribution / packaging
 .Python
