+++ app/main.py
@@ -1,17 +1,17 @@
 from fastapi import FastAPI, File, UploadFile
 from fire_classifier.predictor import ImagePredictor
-from fire_classifier.util.api import API
+from app.settings import API_SETTINGS
 
 # A: init API
 app = FastAPI(
-    title=API["title"],
-    description=API["description"],
-    version=API["version"],
-    openapi_tags=API["endpoints"],
+    title=API_SETTINGS["title"],
+    description=API_SETTINGS["description"],
+    version=API_SETTINGS["version"],
+    openapi_tags=API_SETTINGS["openapi_tags"],
 )
 
 # B: init ML inference object, and the routes
-predictor_config_path = API["config_path"]
+predictor_config_path = API_SETTINGS["predictor_config_path"]
 predictor = ImagePredictor.init_from_config_path(predictor_config_path)
