+++ app/settings.py
@@ -10,7 +10,7 @@ API_VERSION = "0.0.2"
 # These are intended shown on the UI, related only to a specific endpoint.
 # Note: the value in the "name" field should match what goes in the
 #       "tags" parameter of the corresponding app route in main.py!!
-API_ENDPOINTS = (
+API_ENDPOINT_DATA = (
     {
         "name": "Detect Fire",
         "description": "Predicts the possibility that a color image contains fire.",
@@ -18,14 +18,14 @@ API_ENDPOINTS = (
 )
 
 # Tells the app how to find the config.yaml (for running ML inference)
-BASE_DIR = Path(__file__).resolve().parent.parent.parent
+BASE_DIR = Path(__file__).resolve().parent.parent
 CONFIG_PATH = os.path.join(BASE_DIR, "app", "config.yaml")
 
 # Wraps all the API metadata as one dictionary
-API = {
+API_SETTINGS = {
     "title": API_TITLE,
     "description": API_DESCRIPTION,
     "version": API_VERSION,
-    "endpoints": API_ENDPOINTS,
-    "config_path": CONFIG_PATH,
+    "openapi_tags": API_ENDPOINT_DATA,
+    "predictor_config_path": CONFIG_PATH,
 }
