+++ fire_classifier/util/preprocessing.py
